+++ galaxy_ng/app/__init__.py
@@ -1,5 +1,4 @@
 from pulpcore.plugin import PulpPluginAppConfig
-from django.db.models.signals import post_migrate
 
 
 class PulpGalaxyPluginAppConfig(PulpPluginAppConfig):
@@ -11,23 +10,4 @@ class PulpGalaxyPluginAppConfig(PulpPluginAppConfig):
 
     def ready(self):
         super().ready()
-        post_migrate.connect(
-            set_pulp_container_access_policies,
-            sender=self,
-            dispatch_uid="override_pulp_container_access_policies"
-        )
-
-
-def set_pulp_container_access_policies(sender, **kwargs):
-    apps = kwargs.get("apps")
-    if apps is None:
-        from django.apps import apps
-    AccessPolicy = apps.get_model("core", "AccessPolicy")
-
-    from galaxy_ng.app.access_control import statements # noqa
-    PULP_CONTAINER_VIEWSETS = statements.PULP_CONTAINER_VIEWSETS
-
-    print("Overriding pulp_container access poliicy")
-    for view in PULP_CONTAINER_VIEWSETS:
-        policy, created = AccessPolicy.objects.update_or_create(
-            viewset_name=view, defaults={**PULP_CONTAINER_VIEWSETS[view], "customized": True})
+        from .signals import handlers  # noqa
